requirements.txt
numpy
torch==2.8.0
pybind11
python-chess
pytest
pyyaml
matplotlib
seaborn
pandas
ruff
black
mypy